requirements.txt
pdfminer.six==20221105
spacy==3.7.2
scikit-learn
